Refactored Docs structure completed
