+++ src/components/DocsView/components/FAQ.tsx
@@ -3,13 +3,12 @@ import * as cx from 'classnames'
 import { Link } from 'react-router'
 import { $p, $g, Icon, $v } from 'graphcool-styles'
 import styled from 'styled-components'
-import Footer from '../../../Footer/Footer'
-import { breakpoints, maxWidth } from '../../../../utils/constants'
+import Footer from '../../Footer/Footer'
+import { breakpoints, maxWidth } from '../../../utils/constants'
 
 export default class FAQ extends React.Component<{}, {}> {
-  render() {
-    const assetsPath = '../../../../../assets/graphics/'
 
+  render() {
     const RightSection = styled.div`
       flex: 0 0 400px;
       background-color: rgba(0, 0, 0, 0.02);
@@ -45,7 +44,7 @@ export default class FAQ extends React.Component<{}, {}> {
             <div className={cx($p.flex, $p.pb96)}>
               <NavigationLinks className={cx($p.mt10, $p.fw6, $p.mr60)}>COMMUNITY</NavigationLinks>
               <NavigationLinks className={cx($p.mt4)}>
-                <img className={cx($p.bbox, $p.db)} src={require(assetsPath + 'APIEndpoints.png')}/>
+                <img className={cx($p.bbox, $p.db)} src={require('../../../assets/graphics/APIEndpoints.png')}/>
               </NavigationLinks>
             </div>
             <div>
@@ -56,12 +55,12 @@ export default class FAQ extends React.Component<{}, {}> {
                   className={cx($p.bbox, $p.db, $p.mr16, $p.mt4)}
                 >
                   <img
-                    src={require(assetsPath + 'QuestionMark.svg')}
+                    src={require('../../../assets/graphics/QuestionMark.svg')}
                     width={25}
                     height={25}
                   />
                 </Circle>
-                  <p className={cx($p.black60, $p.f20, $p.fw4)}>How can I implement my server-side business logic?</p>
+                <p className={cx($p.black60, $p.f20, $p.fw4)}>How can I implement my server-side business logic?</p>
               </div>
               <div className={cx($p.flex, $p.pv10)}>
                 <Circle
@@ -69,7 +68,7 @@ export default class FAQ extends React.Component<{}, {}> {
                   className={cx($p.bbox, $p.db, $p.mr16, $p.mt4)}
                 >
                   <img
-                    src={require(assetsPath + 'QuestionMark.svg')}
+                    src={require('../../../assets/graphics/QuestionMark.svg')}
                     width={25}
                     height={25}
                   />
@@ -82,7 +81,7 @@ export default class FAQ extends React.Component<{}, {}> {
                   className={cx($p.bbox, $p.db, $p.mr16, $p.mt4)}
                 >
                   <img
-                    src={require(assetsPath + 'QuestionMark.svg')}
+                    src={require('../../../assets/graphics/QuestionMark.svg')}
                     width={25}
                     height={25}
                   />
@@ -93,7 +92,7 @@ export default class FAQ extends React.Component<{}, {}> {
                 <div className={cx($p.inlineFlex, $p.bgLightgreen10, $p.pv25)}>
                   <p className={cx($p.lightgreen50, $p.fw6, $p.f16, $p.pl25)}>MORE ABOUT</p>
                   <p className={cx($p.green, $p.fw6, $p.f16, $p.pl6)}>VENDOR LOCK IN</p>
-                  <img src={require(assetsPath + 'info.svg')}
+                  <img src={require('../../../assets/graphics/info.svg')}
                        className={cx($p.bbox, $p.db, $p.pl25)}
                   />
                 </div>
@@ -104,7 +103,7 @@ export default class FAQ extends React.Component<{}, {}> {
                       className={cx($p.bbox, $p.db, $p.mr16, $p.mt4)}
                     >
                       <img
-                        src={require(assetsPath + 'Guide.svg')}
+                        src={require('../../../assets/graphics/Guide.svg')}
                         width={25}
                         height={25}
                       />
@@ -120,7 +119,7 @@ export default class FAQ extends React.Component<{}, {}> {
                       className={cx($p.bbox, $p.db, $p.mr16, $p.mt4)}
                     >
                       <img
-                        src={require(assetsPath + 'Example.svg')}
+                        src={require('../../../assets/graphics/Example.svg')}
                         width={25}
                         height={25}
                       />
@@ -136,7 +135,7 @@ export default class FAQ extends React.Component<{}, {}> {
                       className={cx($p.bbox, $p.db, $p.mr16, $p.mt4)}
                     >
                       <img
-                        src={require(assetsPath + 'Feather.svg')}
+                        src={require('../../../assets/graphics/Feather.svg')}
                         width={25}
                         height={25}
                       />
@@ -155,7 +154,7 @@ export default class FAQ extends React.Component<{}, {}> {
                   className={cx($p.bbox, $p.db, $p.mr16, $p.mt4)}
                 >
                   <img
-                    src={require(assetsPath + 'QuestionMark.svg')}
+                    src={require('../../../assets/graphics/QuestionMark.svg')}
                     width={25}
                     height={25}
                   />
@@ -171,7 +170,7 @@ export default class FAQ extends React.Component<{}, {}> {
                   className={cx($p.bbox, $p.db, $p.mr16, $p.mt4)}
                 >
                   <img
-                    src={require(assetsPath + 'Filters.svg')}
+                    src={require('../../../assets/graphics/Filters.svg')}
                     width={25}
                     height={25}
                   />
