+++ src/components/DocsView/components/Markdown.tsx
@@ -2,7 +2,6 @@ import * as React from 'react'
 import {Node, Parser} from 'commonmark'
 import * as ReactRenderer from 'commonmark-react-renderer'
 import {PrismCode} from 'react-prism'
-const styles: any = require('./Markdown.module.css')
 
 interface Props {
   ast: Node
