+++ src/components/DocsView/components/RelatedContent.tsx
