+++ src/main.tsx
@@ -6,14 +6,14 @@ import HomeView from './components/HomeView/HomeView'
 import PricingView from './components/PricingView/PricingView'
 import AboutView from './components/AboutView/AboutView'
 import DocsView from './components/DocsView/DocsView'
-import FAQ from './components/DocsView/components/FAQ/FAQ'
+import FAQ from './components/DocsView/components/FAQ'
 import * as FastClick from 'fastclick'
 import QuickstartPage from './components/DocsView/pages/quickstart/QuickstartPage'
 import ReferencePage from './components/DocsView/pages/reference/ReferencePage'
 import ResourcesPage from './components/DocsView/pages/resources/ResourcesPage'
 import BlogPage from './components/DocsView/pages/blog/BlogPage'
 import CommunityPage from './components/DocsView/pages/community/CommunityPage'
-import ContentHandler from './components/DocsView/ContentHandler/ContentHandler'
+import ContentHandler from './components/DocsView/components/ContentHandler'
 import ApolloClient, {createNetworkInterface} from 'apollo-client'
 import { ApolloProvider } from 'react-apollo'
